requirements.txt
django>=3.0
djangorestframework>=3.11
djangorestframework-jwt>=1.11
django-redis
requests